PoliceVideoAISummary/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
celery
redis
python-multipart 